Get the App.test.tsx test to work
